Updating hosting to break up single file into lots of independent commands - this was an attempt to improve syntax. Adding in single image handling, as well as full support for watch directory.
