+++ packages/cli/src/constants.js
@@ -11,7 +11,6 @@ exports.CommandWords = {
 exports.InitTypes = {
   QUICK: "quick",
   DIGITAL_OCEAN: "do",
-  SINGLE: "single",
 }
 
 exports.AnalyticsEvents = {
